requirements.txt
Flask==2.3.3
python-telegram-bot==20.7
google-generativeai>=0.8.0
aiohttp>=3.9.0
